Refactoring App cmponent, separated routing and NavigationAwareComponent
