+++ tnoodle-ui/src/index.js
@@ -1,5 +1,19 @@
 import React from 'react';
+import history from './history';
 import { render } from 'react-dom';
-import { App } from '../src/components/App';
+import getStore from './redux/store';
+import { Provider } from 'react-redux';
+import App from '../src/components/App';
+import { ConnectedRouter } from 'react-router-redux';
 
-render(<App />, document.getElementById('root'));
+const store = getStore({}, history);
+
+const app = (
+  <Provider store={store}>
+    <ConnectedRouter history={history}>
+      <App />
+    </ConnectedRouter>
+  </Provider>
+);
+
+render(app, document.getElementById('root'));
